+++ src/test/java/TestHelper.java
@@ -5,6 +5,7 @@ import java.util.List;
 public class TestHelper {
     public static final Path PATH_TO_SAMPLE_MAVEN_PROJECT =
             Paths.get("src/test/resources/sample-maven-project");
+    public static final Path PATH_TO_INPUT = PATH_TO_SAMPLE_MAVEN_PROJECT.resolve("inputs");
 
     public static String[] getMavenClasspathFromBuildDirectory(Path buildDirectory) {
         List<String> classpath =
