--- src/test/resources/sample-maven-project/collection-input.txt
@@ -1,1 +1,0 @@
-foo.PrintingCollections=12
