--- src/test/resources/sample-maven-project/input.txt
@@ -1,1 +1,0 @@
-foo.BasicMath=5,9
