+++ src/test/resources/sample-maven-project/inputs/basic-math.txt
@@ -1,0 +1,1 @@
+foo.BasicMath=7,11
