+++ src/test/resources/sample-maven-project/inputs/collections/one-level.txt
@@ -1,0 +1,1 @@
+foo.collections.OneLevelCollections=12
