+++ src/test/resources/sample-maven-project/inputs/static-class-field.txt
