+++ src/test/resources/sample-maven-project/src/test/java/foo/CollectionsTest.java
@@ -1,0 +1,20 @@
+package foo;
+
+import static org.junit.jupiter.api.Assertions.assertFalse;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import org.junit.jupiter.api.Test;
+import foo.collections.OneLevelCollections;
+import foo.collections.NestedCollections;
+
+public class CollectionsTest {
+    @Test
+    void test_returnTruthy() {
+        assertTrue(OneLevelCollections.returnTruthy());
+    }
+
+    @Test
+    void test_returnFalsy() {
+        assertFalse(NestedCollections.returnFalsy());
+    }
+}
